Bug 1091008 - Reimplement HasLowUndecodedData in terms of GetBuffered. r=cpearce
Since GetBuffered now has a sane implementation for MSE, this should
make this function sane for MSE as well.

UltraBlame original commit: deae19d73b34feb4547e641eb90f7749d30aa93b

+++ dom/media/MediaDecoderStateMachine.h
@@ -453,7 +453,7 @@ protected:
   bool HasLowUndecodedData();
 
   
-  bool HasLowUndecodedData(double aUsecs);
+  bool HasLowUndecodedData(int64_t aUsecs);
